pyproject.toml
[project]
name = "translate-videos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepmultilingualpunctuation>=1.0.1",
    "faster-whisper>=1.2.0",
    "ffmpeg-python>=0.2.0",
    "ipaddress>=1.0.23",
    "langid>=1.1.6",
    "librosa>=0.11.0",
    "numpy>=2.2.6",
    "openai-whisper>=20250625",
    "pyworld>=0.3.5",
    "pyyaml>=6.0.3",
    "regex>=2025.9.18",
    "sacremoses>=0.1.1",
    "scipy>=1.16.2",
    "sentencepiece>=0.2.1",
    "soundfile>=0.13.1",
    "torch>=2.8.0",
    "transformers>=4.57.1",
    "whisperx>=3.7.4",
]
